+++ article/repository/mysql/mysqlarticle_test.go
@@ -1,4 +1,4 @@
-package repository_test
+package mysql_test
 
 import (
 	"context"
@@ -9,7 +9,8 @@ import (
 	"github.com/stretchr/testify/require"
 	sqlmock "gopkg.in/DATA-DOG/go-sqlmock.v1"
 
-	articleRepo "github.com/bxcodec/go-clean-arch/article/repository"
+	"github.com/bxcodec/go-clean-arch/article/repository"
+	articleMysqlRepo "github.com/bxcodec/go-clean-arch/article/repository/mysql"
 	"github.com/bxcodec/go-clean-arch/domain"
 )
 
@@ -44,8 +45,8 @@ func TestFetch(t *testing.T) {
 	query := "SELECT id,title,content, author_id, updated_at, created_at FROM article WHERE created_at > \\? ORDER BY created_at LIMIT \\?"
 
 	mock.ExpectQuery(query).WillReturnRows(rows)
-	a := articleRepo.NewMysqlArticleRepository(db)
-	cursor := articleRepo.EncodeCursor(mockArticles[1].CreatedAt)
+	a := articleMysqlRepo.NewMysqlArticleRepository(db)
+	cursor := repository.EncodeCursor(mockArticles[1].CreatedAt)
 	num := int64(2)
 	list, nextCursor, err := a.Fetch(context.TODO(), cursor, num)
 	assert.NotEmpty(t, nextCursor)
@@ -70,7 +71,7 @@ func TestGetByID(t *testing.T) {
 	query := "SELECT id,title,content, author_id, updated_at, created_at FROM article WHERE ID = \\?"
 
 	mock.ExpectQuery(query).WillReturnRows(rows)
-	a := articleRepo.NewMysqlArticleRepository(db)
+	a := articleMysqlRepo.NewMysqlArticleRepository(db)
 
 	num := int64(5)
 	anArticle, err := a.GetByID(context.TODO(), num)
@@ -103,7 +104,7 @@ func TestStore(t *testing.T) {
 	prep := mock.ExpectPrepare(query)
 	prep.ExpectExec().WithArgs(ar.Title, ar.Content, ar.Author.ID, ar.CreatedAt, ar.UpdatedAt).WillReturnResult(sqlmock.NewResult(12, 1))
 
-	a := articleRepo.NewMysqlArticleRepository(db)
+	a := articleMysqlRepo.NewMysqlArticleRepository(db)
 
 	err = a.Store(context.TODO(), ar)
 	assert.NoError(t, err)
@@ -125,7 +126,7 @@ func TestGetByTitle(t *testing.T) {
 	query := "SELECT id,title,content, author_id, updated_at, created_at FROM article WHERE title = \\?"
 
 	mock.ExpectQuery(query).WillReturnRows(rows)
-	a := articleRepo.NewMysqlArticleRepository(db)
+	a := articleMysqlRepo.NewMysqlArticleRepository(db)
 
 	title := "title 1"
 	anArticle, err := a.GetByTitle(context.TODO(), title)
@@ -148,7 +149,7 @@ func TestDelete(t *testing.T) {
 	prep := mock.ExpectPrepare(query)
 	prep.ExpectExec().WithArgs(12).WillReturnResult(sqlmock.NewResult(12, 1))
 
-	a := articleRepo.NewMysqlArticleRepository(db)
+	a := articleMysqlRepo.NewMysqlArticleRepository(db)
 
 	num := int64(12)
 	err = a.Delete(context.TODO(), num)
@@ -183,7 +184,7 @@ func TestUpdate(t *testing.T) {
 	prep := mock.ExpectPrepare(query)
 	prep.ExpectExec().WithArgs(ar.Title, ar.Content, ar.Author.ID, ar.UpdatedAt, ar.ID).WillReturnResult(sqlmock.NewResult(12, 1))
 
-	a := articleRepo.NewMysqlArticleRepository(db)
+	a := articleMysqlRepo.NewMysqlArticleRepository(db)
 
 	err = a.Update(context.TODO(), ar)
 	assert.NoError(t, err)
