+++ author/repository/mysql/mysql_test.go
@@ -1,4 +1,4 @@
-package repository_test
+package mysql_test
 
 import (
 	"context"
@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/require"
 	sqlmock "gopkg.in/DATA-DOG/go-sqlmock.v1"
 
-	"github.com/bxcodec/go-clean-arch/author/repository"
+	repository "github.com/bxcodec/go-clean-arch/author/repository/mysql"
 )
 
 func TestGetByID(t *testing.T) {
